@@ -1,242 +1,153 @@
-package com.zane.smapiinstaller.logic;
+package com.zane.smapiinstaller.logic
 
-import android.app.Activity;
-import android.content.pm.PackageInfo;
-import android.net.Uri;
-import android.util.Log;
-import android.view.View;
-import android.widget.Toast;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.google.common.base.Joiner;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableListMultimap;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Multimaps;
-import com.google.common.collect.Queues;
-import com.lzy.okgo.OkGo;
-import com.lzy.okgo.model.Response;
-import com.microsoft.appcenter.crashes.Crashes;
-import com.zane.smapiinstaller.MobileNavigationDirections;
-import com.zane.smapiinstaller.R;
-import com.zane.smapiinstaller.constant.Constants;
-import com.zane.smapiinstaller.constant.DialogAction;
-import com.zane.smapiinstaller.dto.ModUpdateCheckRequestDto;
-import com.zane.smapiinstaller.dto.ModUpdateCheckResponseDto;
-import com.zane.smapiinstaller.dto.Tuple2;
-import com.zane.smapiinstaller.entity.ModManifestEntry;
-import com.zane.smapiinstaller.utils.DialogUtils;
-import com.zane.smapiinstaller.utils.FileUtils;
-import com.zane.smapiinstaller.utils.JsonCallback;
-import com.zane.smapiinstaller.utils.JsonUtil;
-import com.zane.smapiinstaller.utils.VersionUtil;
-
-import org.apache.commons.lang3.StringUtils;
-import org.zeroturnaround.zip.ZipUtil;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Objects;
-import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Consumer;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
-
-import androidx.documentfile.provider.DocumentFile;
-import androidx.documentfile.provider.DocumentUtils;
-import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
+import android.content.pm.PackageInfo
+import android.util.Log
+import android.view.View
+import android.widget.Toast
+import androidx.documentfile.provider.DocumentFile
+import androidx.documentfile.provider.DocumentUtils
+import androidx.navigation.Navigation.findNavController
+import com.fasterxml.jackson.core.type.TypeReference
+import com.google.common.base.Joiner
+import com.google.common.collect.ImmutableList
+import com.google.common.collect.ImmutableListMultimap
+import com.google.common.collect.Lists
+import com.google.common.collect.Multimaps
+import com.google.common.collect.Queues
+import com.lzy.okgo.OkGo
+import com.lzy.okgo.model.Response
+import com.microsoft.appcenter.crashes.Crashes
+import com.zane.smapiinstaller.MobileNavigationDirections
+import com.zane.smapiinstaller.R
+import com.zane.smapiinstaller.constant.Constants
+import com.zane.smapiinstaller.constant.DialogAction
+import com.zane.smapiinstaller.dto.ModUpdateCheckRequestDto
+import com.zane.smapiinstaller.dto.ModUpdateCheckRequestDto.ModInfo
+import com.zane.smapiinstaller.dto.ModUpdateCheckRequestDto.SemanticVersion
+import com.zane.smapiinstaller.dto.ModUpdateCheckResponseDto
+import com.zane.smapiinstaller.dto.ModUpdateCheckResponseDto.UpdateInfo
+import com.zane.smapiinstaller.dto.Tuple2
+import com.zane.smapiinstaller.entity.ModManifestEntry
+import com.zane.smapiinstaller.logic.CommonLogic.checkDataRootPermission
+import com.zane.smapiinstaller.logic.CommonLogic.copyDocument
+import com.zane.smapiinstaller.logic.CommonLogic.getActivityFromView
+import com.zane.smapiinstaller.logic.CommonLogic.pathToTreeUri
+import com.zane.smapiinstaller.utils.DialogUtils
+import com.zane.smapiinstaller.utils.DialogUtils.showAlertDialog
+import com.zane.smapiinstaller.utils.FileUtils
+import com.zane.smapiinstaller.utils.JsonCallback
+import com.zane.smapiinstaller.utils.JsonUtil
+import com.zane.smapiinstaller.utils.VersionUtil
+import org.apache.commons.lang3.StringUtils
+import org.zeroturnaround.zip.ZipUtil
+import java.io.File
+import java.io.IOException
+import java.util.concurrent.atomic.AtomicBoolean
 
 /**
  * Mod资源管理器
  *
  * @author Zane
  */
-public class ModAssetsManager {
-
-    private final View root;
-
-    private static final String TAG = "MANAGER";
-
-    private final AtomicBoolean checkUpdating = new AtomicBoolean(false);
-
-    public ModAssetsManager(View root) {
-        this.root = root;
-    }
-
-    /**
-     * 查找第一个匹配的Mod
-     *
-     * @param filter 过滤规则
-     * @return Mod信息
-     */
-    public static ModManifestEntry findFirstModIf(Predicate<ModManifestEntry> filter) {
-        ConcurrentLinkedQueue<File> files = Queues.newConcurrentLinkedQueue();
-        files.add(new File(FileUtils.getStadewValleyBasePath(), Constants.MOD_PATH));
-        do {
-            File currentFile = files.poll();
-            if (currentFile != null && currentFile.exists()) {
-                boolean foundManifest = false;
-                File[] listFiles = currentFile.listFiles(File::isFile);
-                if (listFiles == null) {
-                    continue;
-                }
-                for (File file : listFiles) {
-                    if (StringUtils.equalsIgnoreCase(file.getName(), "manifest.json")) {
-                        ModManifestEntry manifest = FileUtils.getFileJson(file, ModManifestEntry.class);
-                        foundManifest = true;
-                        if (manifest != null) {
-                            manifest.setAssetPath(file.getParentFile().getAbsolutePath());
-                            if (filter.test(manifest)) {
-                                return manifest;
-                            }
-                        }
-                        break;
-                    }
-                }
-                if (!foundManifest) {
-                    files.addAll(Lists.newArrayList(currentFile.listFiles(File::isDirectory)));
-                }
-            }
-        } while (!files.isEmpty());
-        return null;
-    }
-
-    /**
-     * 查找全部已识别Mod
-     *
-     * @return Mod信息列表
-     */
-    public static List<ModManifestEntry> findAllInstalledMods() {
-        return findAllInstalledMods(false);
-    }
-
-    /**
-     * 查找全部已识别Mod
-     *
-     * @param ignoreDisabledMod 是否忽略禁用的mod
-     * @return Mod信息列表
-     */
-    public static List<ModManifestEntry> findAllInstalledMods(boolean ignoreDisabledMod) {
-        ConcurrentLinkedQueue<File> files = Queues.newConcurrentLinkedQueue();
-        files.add(new File(FileUtils.getStadewValleyBasePath(), Constants.MOD_PATH));
-        List<ModManifestEntry> mods = new ArrayList<>(30);
-        do {
-            File currentFile = files.poll();
-            if (currentFile != null && currentFile.exists()) {
-                boolean foundManifest = false;
-                File[] listFiles = currentFile.listFiles(File::isFile);
-                if (listFiles != null) {
-                    for (File file : listFiles) {
-                        if (StringUtils.equalsIgnoreCase(file.getName(), "manifest.json")) {
-                            ModManifestEntry manifest = FileUtils.getFileJson(file, ModManifestEntry.class);
-                            foundManifest = true;
-                            if (manifest != null && StringUtils.isNoneBlank(manifest.getUniqueID())) {
-                                if (ignoreDisabledMod && StringUtils.startsWith(file.getParentFile().getName(), ".")) {
-                                    break;
-                                }
-                                manifest.setAssetPath(file.getParentFile().getAbsolutePath());
-                                manifest.setLastModified(file.lastModified());
-                                mods.add(manifest);
-                            }
-                            break;
-                        }
-                    }
-                }
-                if (!foundManifest) {
-                    File[] listDirectories = currentFile.listFiles(File::isDirectory);
-                    if (listDirectories != null) {
-                        files.addAll(Lists.newArrayList(listDirectories));
-                    }
-                }
-            }
-        } while (!files.isEmpty());
-        return mods;
-    }
+class ModAssetsManager(private val root: View) {
+    private val checkUpdating = AtomicBoolean(false)
 
     /**
      * 安装默认Mod
      *
      * @return 是否安装成功
      */
-    public boolean installDefaultMods() {
-        Activity context = CommonLogic.getActivityFromView(root);
-        List<ModManifestEntry> modManifestEntries = FileUtils.getAssetJson(context, "mods_manifest.json", new TypeReference<List<ModManifestEntry>>() {
-        });
-        if (modManifestEntries == null) {
-            return false;
-        }
-        File modFolder = new File(FileUtils.getStadewValleyBasePath(), Constants.MOD_PATH);
-        ImmutableListMultimap<String, ModManifestEntry> installedModMap = Multimaps.index(findAllInstalledMods(), ModManifestEntry::getUniqueID);
-        List<File> unpackedMods = new ArrayList<>();
-        for (ModManifestEntry mod : modManifestEntries) {
-            if (installedModMap.containsKey(mod.getUniqueID())) {
-                ImmutableList<ModManifestEntry> installedMods = installedModMap.get(mod.getUniqueID());
-                if (installedMods.size() > 1) {
-                    DialogUtils.showAlertDialog(root, R.string.error,
-                            String.format(context.getString(R.string.duplicate_mod_found),
-                                    installedMods.stream().map(item -> FileUtils.toPrettyPath(item.getAssetPath())).collect(Collectors.joining(","))));
-                    return false;
+    fun installDefaultMods(): Boolean {
+        val context = getActivityFromView(root)
+        val modManifestEntries: List<ModManifestEntry> = context?.let {
+            FileUtils.getAssetJson(it,
+                "mods_manifest.json",
+                object : TypeReference<List<ModManifestEntry>>() {})
+        } ?: return false
+        val modFolder = File(FileUtils.stadewValleyBasePath, Constants.MOD_PATH)
+        val installedModMap = Multimaps.index(findAllInstalledMods()) { obj -> obj!!.uniqueID }
+        val unpackedMods: MutableList<File> = ArrayList()
+        for (mod in modManifestEntries) {
+            if (installedModMap.containsKey(mod.uniqueID)) {
+                val installedMods = installedModMap[mod.uniqueID]
+                if (installedMods.size > 1) {
+                    showAlertDialog(
+                        root,
+                        R.string.error,
+                        String.format(context.getString(R.string.duplicate_mod_found),
+                            installedMods.joinToString(",") { item ->
+                                FileUtils.toPrettyPath(item.assetPath)
+                            })
+                    )
+                    return false
                 }
-                if (mod.getCleanInstall() != null && mod.getCleanInstall()) {
+                if (mod.cleanInstall != null && mod.cleanInstall) {
                     // Delete origin version with different Unique ID
-                    if(mod.getOriginUniqueId() != null) {
-                        installedModMap.keys().stream().filter(id -> mod.getOriginUniqueId().contains(id)).map(installedModMap::get).forEach(list -> {
-                            for (ModManifestEntry entry : list) {
-                                try {
-                                    FileUtils.deleteDirectory(new File(entry.getAssetPath()));
-                                } catch (IOException e) {
-                                    Log.e(TAG, "Install Mod Error", e);
+                    if (mod.originUniqueId != null) {
+                        installedModMap.keys().filter { id -> mod.originUniqueId.contains(id) }
+                            .map { key -> installedModMap[key] }.forEach { list ->
+                                for (entry in list) {
+                                    try {
+                                        org.zeroturnaround.zip.commons.FileUtils.deleteDirectory(
+                                            File(entry.assetPath)
+                                        )
+                                    } catch (e: IOException) {
+                                        Log.e(TAG, "Install Mod Error", e)
+                                    }
                                 }
                             }
-                        });
                     }
                     // Delete old version
-                    if(installedMods.size() > 0 && mod.getVersion() != null && VersionUtil.compareVersion(installedMods.get(0).getVersion(), mod.getVersion()) < 0) {
+                    if (installedMods.size > 0 && mod.version != null && VersionUtil.compareVersion(
+                            installedMods[0].version, mod.version
+                        ) < 0
+                    ) {
                         try {
-                            FileUtils.deleteDirectory(new File(installedMods.get(0).getAssetPath()));
-                        } catch (IOException e) {
-                            Log.e(TAG, "Install Mod Error", e);
+                            org.zeroturnaround.zip.commons.FileUtils.deleteDirectory(
+                                File(
+                                    installedMods[0].assetPath
+                                )
+                            )
+                        } catch (e: IOException) {
+                            Log.e(TAG, "Install Mod Error", e)
                         }
                     }
                 }
-                if (installedMods.size() > 0) {
+                if (installedMods.size > 0) {
                     try {
-                        File targetFile = new File(installedMods.get(0).getAssetPath());
-                        ZipUtil.unpack(context.getAssets().open(mod.getAssetPath()), targetFile, (name) -> StringUtils.removeStart(name, mod.getName() + "/"));
-                        unpackedMods.add(targetFile);
-                    } catch (IOException e) {
-                        Log.e(TAG, "Install Mod Error", e);
+                        val targetFile = File(installedMods[0].assetPath)
+                        ZipUtil.unpack(
+                            context.assets.open(mod.assetPath), targetFile
+                        ) { name -> StringUtils.removeStart(name, mod.name + "/") }
+                        unpackedMods.add(targetFile)
+                    } catch (e: IOException) {
+                        Log.e(TAG, "Install Mod Error", e)
                     }
-                    continue;
+                    continue
                 }
             }
             try {
-                ZipUtil.unpack(context.getAssets().open(mod.getAssetPath()), modFolder);
-                unpackedMods.add(new File(modFolder, mod.getName()));
-            } catch (IOException e) {
-                Log.e(TAG, "Install Mod Error", e);
+                ZipUtil.unpack(context.assets.open(mod.assetPath), modFolder)
+                unpackedMods.add(File(modFolder, mod.name))
+            } catch (e: IOException) {
+                Log.e(TAG, "Install Mod Error", e)
             }
         }
-        if (CommonLogic.checkDataRootPermission(context)) {
-            Uri targetDirUri = CommonLogic.pathToTreeUri(Constants.TARGET_DATA_FILE_URI);
-            DocumentFile documentFile = DocumentFile.fromTreeUri(context, targetDirUri);
-            if(documentFile != null) {
-                DocumentFile filesDoc = DocumentUtils.findFile(context, documentFile, "files");
-                DocumentFile modsDoc = DocumentUtils.findFile(context, filesDoc, "Mods");
-                if (modsDoc == null) {
-                    modsDoc = filesDoc.createDirectory("Mods");
-                }
-                for (File mod : unpackedMods) {
-                    CommonLogic.copyDocument(context, mod, modsDoc);
+        if (checkDataRootPermission(context)) {
+            val targetDirUri = pathToTreeUri(Constants.TARGET_DATA_FILE_URI)
+            val documentFile = DocumentFile.fromTreeUri(context, targetDirUri)
+            if (documentFile != null) {
+                val filesDoc = DocumentUtils.findFile(context, documentFile, "files")
+                val modsDoc = DocumentUtils.findFile(context, filesDoc, "Mods")
+                    ?: filesDoc.createDirectory("Mods")
+                modsDoc?.let {
+                    for (mod in unpackedMods) {
+                        copyDocument(context, mod, modsDoc)
+                    }
                 }
             }
         }
-        return true;
+        return true
     }
 
     /**
@@ -244,21 +155,21 @@ public class ModAssetsManager {
      *
      * @param returnCallback 回调函数
      */
-    public void checkModEnvironment(Consumer<Boolean> returnCallback) {
-        ImmutableListMultimap<String, ModManifestEntry> installedModMap = Multimaps.index(findAllInstalledMods(true), ModManifestEntry::getUniqueID);
-        checkDuplicateMod(installedModMap, (isConfirm) -> {
+    fun checkModEnvironment(returnCallback: (Boolean) -> Unit) {
+        val installedModMap = Multimaps.index(findAllInstalledMods(true)) { obj -> obj!!.uniqueID }
+        checkDuplicateMod(installedModMap) { isConfirm ->
             if (isConfirm) {
-                checkUnsatisfiedDependencies(installedModMap, (isConfirm2) -> {
+                checkUnsatisfiedDependencies(installedModMap) { isConfirm2 ->
                     if (isConfirm2) {
-                        checkContentpacks(installedModMap, returnCallback);
+                        checkContentpacks(installedModMap, returnCallback)
                     } else {
-                        returnCallback.accept(false);
+                        returnCallback.invoke(false)
                     }
-                });
+                }
             } else {
-                returnCallback.accept(false);
+                returnCallback.invoke(false)
             }
-        });
+        }
     }
 
     /**
@@ -267,22 +178,28 @@ public class ModAssetsManager {
      * @param installedModMap 已安装Mod集合
      * @param returnCallback  回调函数
      */
-    private void checkDuplicateMod(ImmutableListMultimap<String, ModManifestEntry> installedModMap, Consumer<Boolean> returnCallback) {
+    private fun checkDuplicateMod(
+        installedModMap: ImmutableListMultimap<String, ModManifestEntry>,
+        returnCallback: (Boolean) -> Unit
+    ) {
         // Duplicate mod check
-        ArrayList<String> list = Lists.newArrayList();
-        for (String key : installedModMap.keySet()) {
-            ImmutableList<ModManifestEntry> installedMods = installedModMap.get(key);
-            if (installedMods.size() > 1) {
-                list.add(installedMods.stream().map(item -> FileUtils.toPrettyPath(item.getAssetPath())).collect(Collectors.joining(",")));
+        val list = Lists.newArrayList<String?>()
+        for (key in installedModMap.keySet()) {
+            val installedMods = installedModMap[key]
+            if (installedMods.size > 1) {
+                list.add(installedMods.joinToString(",") { item -> FileUtils.toPrettyPath(item.assetPath) })
             }
         }
         if (!list.isEmpty()) {
-            DialogUtils.showConfirmDialog(root, R.string.error,
-                    root.getContext().getString(R.string.duplicate_mod_found, Joiner.on(";").join(list)),
-                    R.string.continue_text, R.string.abort,
-                    ((dialog, which) -> returnCallback.accept(which == DialogAction.POSITIVE)));
+            DialogUtils.showConfirmDialog(
+                root,
+                R.string.error,
+                root.context.getString(R.string.duplicate_mod_found, Joiner.on(";").join(list)),
+                R.string.continue_text,
+                R.string.abort
+            ) { _, which -> returnCallback.invoke(which === DialogAction.POSITIVE) }
         } else {
-            returnCallback.accept(true);
+            returnCallback.invoke(true)
         }
     }
 
@@ -292,34 +209,36 @@ public class ModAssetsManager {
      * @param installedModMap 已安装Mod集合
      * @param returnCallback  回调函数
      */
-    private void checkUnsatisfiedDependencies(ImmutableListMultimap<String, ModManifestEntry> installedModMap, Consumer<Boolean> returnCallback) {
-        List<Tuple2<String, List<String>>> dependencyErrors = installedModMap.values().stream()
-                .map(mod -> checkModDependencyError(mod, installedModMap))
-                .filter(Objects::nonNull)
-                .collect(Collectors.toList());
-        if (dependencyErrors.size() > 0) {
-            DialogUtils.showConfirmDialog(root, R.string.error,
-                    dependencyErrors.stream().map(Tuple2::getFirst).collect(Collectors.joining(";")),
-                    R.string.continue_text, R.string.menu_download,
-                    ((dialog, which) -> {
-                        if (which == DialogAction.POSITIVE) {
-                            returnCallback.accept(true);
-                        } else {
-                            List<ModUpdateCheckRequestDto.ModInfo> list = dependencyErrors.stream()
-                                    .map(Tuple2::getSecond).flatMap(Collection::stream).distinct()
-                                    .map(item -> {
-                                        ModUpdateCheckRequestDto.ModInfo modInfo = new ModUpdateCheckRequestDto.ModInfo();
-                                        modInfo.setId(item);
-                                        return modInfo;
-                                    })
-                                    .distinct()
-                                    .collect(Collectors.toList());
-                            redirectModDownload(list);
-                            returnCallback.accept(false);
-                        }
-                    }));
+    private fun checkUnsatisfiedDependencies(
+        installedModMap: ImmutableListMultimap<String, ModManifestEntry>,
+        returnCallback: (Boolean) -> Unit
+    ) {
+        val dependencyErrors = installedModMap.values()
+            .mapNotNull { mod -> checkModDependencyError(mod, installedModMap) }.toList()
+        if (dependencyErrors.isNotEmpty()) {
+            DialogUtils.showConfirmDialog(
+                root,
+                R.string.error,
+                dependencyErrors.map { obj -> obj.first }.joinToString(";"),
+                R.string.continue_text,
+                R.string.menu_download
+            ) { _, which ->
+                if (which === DialogAction.POSITIVE) {
+                    returnCallback.invoke(true)
+                } else {
+                    val list =
+                        dependencyErrors.asSequence().map { obj -> obj.second }.flatten().distinct()
+                            .map { item ->
+                                val modInfo = ModInfo()
+                                modInfo.id = item
+                                modInfo
+                            }.distinct().toList()
+                    redirectModDownload(list)
+                    returnCallback.invoke(false)
+                }
+            }
         } else {
-            returnCallback.accept(true);
+            returnCallback.invoke(true)
         }
     }
 
@@ -329,190 +248,311 @@ public class ModAssetsManager {
      * @param installedModMap 已安装Mod集合
      * @param returnCallback  回调函数
      */
-    private void checkContentpacks(ImmutableListMultimap<String, ModManifestEntry> installedModMap, Consumer<Boolean> returnCallback) {
-        List<Tuple2<String, String>> dependencyErrors = installedModMap.values().stream()
-                .map(mod -> checkContentPackDependencyError(mod, installedModMap))
-                .filter(Objects::nonNull).collect(Collectors.toList());
-        if (!dependencyErrors.isEmpty()) {
-            DialogUtils.showConfirmDialog(root, R.string.error,
-                    dependencyErrors.stream().map(Tuple2::getFirst).collect(Collectors.joining(";")),
-                    R.string.continue_text, R.string.menu_download,
-                    ((dialog, which) -> {
-                        if (which == DialogAction.POSITIVE) {
-                            returnCallback.accept(true);
-                        } else {
-                            List<ModUpdateCheckRequestDto.ModInfo> list = dependencyErrors.stream()
-                                    .map(item -> {
-                                        ModUpdateCheckRequestDto.ModInfo modInfo = new ModUpdateCheckRequestDto.ModInfo();
-                                        modInfo.setId(item.getSecond());
-                                        return modInfo;
-                                    })
-                                    .distinct()
-                                    .collect(Collectors.toList());
-                            redirectModDownload(list);
-                            returnCallback.accept(false);
-                        }
-                    }));
+    private fun checkContentpacks(
+        installedModMap: ImmutableListMultimap<String, ModManifestEntry>,
+        returnCallback: (Boolean) -> Unit
+    ) {
+        val dependencyErrors = installedModMap.values().mapNotNull { mod ->
+            checkContentPackDependencyError(
+                mod, installedModMap
+            )
+        }.toList()
+        if (dependencyErrors.isNotEmpty()) {
+            DialogUtils.showConfirmDialog(
+                root,
+                R.string.error,
+                dependencyErrors.map { obj -> obj.first }.joinToString(";"),
+                R.string.continue_text,
+                R.string.menu_download
+            ) { _, which ->
+                if (which === DialogAction.POSITIVE) {
+                    returnCallback.invoke(true)
+                } else {
+                    val list = dependencyErrors.map { item ->
+                        val modInfo = ModInfo()
+                        modInfo.id = item.second
+                        modInfo
+                    }.distinct().toList()
+                    redirectModDownload(list)
+                    returnCallback.invoke(false)
+                }
+            }
         } else {
-            returnCallback.accept(true);
+            returnCallback.invoke(true)
         }
     }
 
-    private void redirectModDownload(List<ModUpdateCheckRequestDto.ModInfo> list) {
-        ModUpdateCheckRequestDto requestDto = new ModUpdateCheckRequestDto(list);
+    private fun redirectModDownload(list: List<ModInfo>) {
+        val requestDto = ModUpdateCheckRequestDto(list)
         try {
-            requestDto.setIncludeExtendedMetadata(true);
-            OkGo.<List<ModUpdateCheckResponseDto>>post(Constants.UPDATE_CHECK_SERVICE_URL)
-                    .upJson(JsonUtil.toJson(requestDto))
-                    .execute(new JsonCallback<List<ModUpdateCheckResponseDto>>(new TypeReference<List<ModUpdateCheckResponseDto>>() {
-                    }) {
-                        @Override
-                        public void onError(Response<List<ModUpdateCheckResponseDto>> response) {
-                            super.onError(response);
-                        }
+            requestDto.isIncludeExtendedMetadata = true
+            OkGo.post<List<ModUpdateCheckResponseDto>>(Constants.UPDATE_CHECK_SERVICE_URL)
+                .upJson(JsonUtil.toJson(requestDto)).execute(object :
+                    JsonCallback<List<ModUpdateCheckResponseDto>>(object :
+                        TypeReference<List<ModUpdateCheckResponseDto>>() {}) {
+                    override fun onError(response: Response<List<ModUpdateCheckResponseDto>>) {
+                        super.onError(response)
+                    }
 
-                        @Override
-                        public void onSuccess(Response<List<ModUpdateCheckResponseDto>> response) {
-                            List<ModUpdateCheckResponseDto> checkResponseDtos = response.body();
-                            if (checkResponseDtos != null) {
-                                List<ModUpdateCheckResponseDto> list = checkResponseDtos.stream()
-                                        .filter(dto -> dto.getMetadata() != null && dto.getMetadata().getMain() != null && StringUtils.isNoneBlank(dto.getMetadata().getMain().getUrl()))
-                                        .collect(Collectors.toList());
-                                try {
-                                    if (list.size() > 0) {
-                                        for (ModUpdateCheckResponseDto dto : list) {
-                                            dto.setSuggestedUpdate(new ModUpdateCheckResponseDto.UpdateInfo(dto.getMetadata().getMain().getVersion(), dto.getMetadata().getMain().getUrl()));
-                                        }
-                                        NavController controller = Navigation.findNavController(CommonLogic.getActivityFromView(root), R.id.nav_host_fragment);
-                                        controller.navigate(MobileNavigationDirections.actionNavAnyToModUpdateFragment(JsonUtil.toJson(list)));
+                    override fun onSuccess(response: Response<List<ModUpdateCheckResponseDto>>) {
+                        val checkResponseDtos = response.body()
+                        if (checkResponseDtos != null) {
+                            val validList = checkResponseDtos.filter { dto ->
+                                dto.metadata != null && dto.metadata.main != null && StringUtils.isNoneBlank(
+                                    dto.metadata.main.url
+                                )
+                            }.toList()
+                            try {
+                                if (validList.isNotEmpty()) {
+                                    for (dto in validList) {
+                                        dto.suggestedUpdate = UpdateInfo(
+                                            dto.metadata.main.version, dto.metadata.main.url
+                                        )
                                     }
-                                } catch (Exception e) {
-                                    Crashes.trackError(e);
+                                    getActivityFromView(root)?.let {
+                                        val controller =
+                                            findNavController(it, R.id.nav_host_fragment)
+                                        controller.navigate(
+                                            MobileNavigationDirections.actionNavAnyToModUpdateFragment(
+                                                JsonUtil.toJson(validList)
+                                            )
+                                        )
+                                    }
                                 }
+                            } catch (e: Exception) {
+                                Crashes.trackError(e)
                             }
                         }
-                    });
-        } catch (Exception e) {
-            Crashes.trackError(e);
+                    }
+                })
+        } catch (e: Exception) {
+            Crashes.trackError(e)
         }
     }
 
-    public void checkModUpdate(Consumer<List<ModUpdateCheckResponseDto>> callback) {
+    fun checkModUpdate(callback: (List<ModUpdateCheckResponseDto>) -> Unit) {
         if (checkUpdating.get()) {
-            return;
+            return
         }
-        List<ModUpdateCheckRequestDto.ModInfo> list = findAllInstalledMods(false).stream()
-                .filter(mod -> mod.getUpdateKeys() != null && !mod.getUpdateKeys().isEmpty())
-                .map(ModUpdateCheckRequestDto.ModInfo::fromModManifestEntry)
-                .distinct()
-                .filter(modInfo -> modInfo.getInstalledVersion() != null)
-                .collect(Collectors.toList());
-        Activity context = CommonLogic.getActivityFromView(root);
-        PackageInfo gamePackageInfo = GameLauncher.getGamePackageInfo(context);
-        if (gamePackageInfo == null) {
-            return;
+        val list = findAllInstalledMods(false).asSequence()
+            .filter { mod -> mod.updateKeys != null && mod.updateKeys.isNotEmpty() }
+            .map { mod -> ModInfo.fromModManifestEntry(mod) }.distinct()
+            .filter { modInfo -> modInfo.installedVersion != null }.toList()
+        val context = getActivityFromView(root) ?: return
+        val gamePackageInfo: PackageInfo = GameLauncher.getGamePackageInfo(context) ?: return
+        context.runOnUiThread {
+            Toast.makeText(
+                context, R.string.mod_version_update_checking, Toast.LENGTH_SHORT
+            ).show()
         }
-        context.runOnUiThread(() -> Toast.makeText(context, R.string.mod_version_update_checking, Toast.LENGTH_SHORT).show());
-        checkUpdating.set(true);
+        checkUpdating.set(true)
         try {
-            ModUpdateCheckRequestDto requestDto = new ModUpdateCheckRequestDto(list, new ModUpdateCheckRequestDto.SemanticVersion(gamePackageInfo.versionName));
-            OkGo.<List<ModUpdateCheckResponseDto>>post(Constants.UPDATE_CHECK_SERVICE_URL)
-                    .upJson(JsonUtil.toJson(requestDto))
-                    .execute(new JsonCallback<List<ModUpdateCheckResponseDto>>(new TypeReference<List<ModUpdateCheckResponseDto>>() {
-                    }) {
-                        @Override
-                        public void onError(Response<List<ModUpdateCheckResponseDto>> response) {
-                            super.onError(response);
-                            checkUpdating.set(false);
-                        }
+            val requestDto =
+                ModUpdateCheckRequestDto(list, SemanticVersion(gamePackageInfo.versionName))
+            OkGo.post<List<ModUpdateCheckResponseDto>>(Constants.UPDATE_CHECK_SERVICE_URL)
+                .upJson(JsonUtil.toJson(requestDto)).execute(object :
+                    JsonCallback<List<ModUpdateCheckResponseDto>>(object :
+                        TypeReference<List<ModUpdateCheckResponseDto>>() {}) {
+                    override fun onError(response: Response<List<ModUpdateCheckResponseDto>>) {
+                        super.onError(response)
+                        checkUpdating.set(false)
+                    }
 
-                        @Override
-                        public void onSuccess(Response<List<ModUpdateCheckResponseDto>> response) {
-                            checkUpdating.set(false);
-                            List<ModUpdateCheckResponseDto> checkResponseDtos = response.body();
-                            if (checkResponseDtos != null) {
-                                List<ModUpdateCheckResponseDto> list = checkResponseDtos.stream().filter(dto -> dto.getSuggestedUpdate() != null).collect(Collectors.toList());
-                                callback.accept(list);
-                            }
-                        }
-                    });
-        } catch (Exception e) {
-            checkUpdating.set(false);
-            Crashes.trackError(e);
+                    override fun onSuccess(response: Response<List<ModUpdateCheckResponseDto>>) {
+                        checkUpdating.set(false)
+                        val checkResponseDtos = response.body()
+                        checkResponseDtos?.filter { dto -> dto.suggestedUpdate != null }?.toList()
+                            ?.apply(callback)
+                    }
+                })
+        } catch (e: Exception) {
+            checkUpdating.set(false)
+            Crashes.trackError(e)
         }
     }
 
-    private Tuple2<String, List<String>> checkModDependencyError(ModManifestEntry mod, ImmutableListMultimap<String, ModManifestEntry> installedModMap) {
-        if (mod.getDependencies() != null) {
-            List<ModManifestEntry> unsatisfiedDependencies = mod.getDependencies().stream()
-                    .filter(dependency -> isDependencyIsExist(dependency, installedModMap))
-                    .collect(Collectors.toList());
-            if (unsatisfiedDependencies.size() > 0) {
-                return new Tuple2<>(
-                        root.getContext().getString(R.string.error_depends_on_mod, mod.getUniqueID(), Joiner.on(",").join(Lists.transform(unsatisfiedDependencies, ModManifestEntry::getUniqueID))),
-                        unsatisfiedDependencies.stream().map(ModManifestEntry::getUniqueID).collect(Collectors.toList()));
+    private fun checkModDependencyError(
+        mod: ModManifestEntry, installedModMap: ImmutableListMultimap<String, ModManifestEntry>
+    ): Tuple2<String?, List<String?>>? {
+        if (mod.dependencies != null) {
+            val unsatisfiedDependencies = mod.dependencies.filter { dependency ->
+                isDependencyIsExist(
+                    dependency, installedModMap
+                )
+            }.toList()
+            if (unsatisfiedDependencies.isNotEmpty()) {
+                return Tuple2(
+                    root.context.getString(R.string.error_depends_on_mod,
+                        mod.uniqueID,
+                        unsatisfiedDependencies.joinToString(",") { obj -> obj.uniqueID }),
+                    unsatisfiedDependencies.map { obj -> obj.uniqueID }.toList()
+                )
             }
         }
-        return null;
+        return null
     }
 
-    private boolean isDependencyIsExist(ModManifestEntry dependency, ImmutableListMultimap<String, ModManifestEntry> installedModMap) {
-        if (dependency.getIsRequired() != null && !dependency.getIsRequired()) {
-            return false;
+    private fun isDependencyIsExist(
+        dependency: ModManifestEntry,
+        installedModMap: ImmutableListMultimap<String, ModManifestEntry>
+    ): Boolean {
+        if (dependency.isRequired != null && !dependency.isRequired) {
+            return false
         }
-        ImmutableList<ModManifestEntry> entries = installedModMap.get(dependency.getUniqueID());
-        if (entries.size() == 0) {
-            for (String key : installedModMap.keySet()) {
-                if (StringUtils.equalsIgnoreCase(key, dependency.getUniqueID())) {
-                    dependency.setUniqueID(key);
-                    entries = installedModMap.get(dependency.getUniqueID());
-                    break;
+        var entries = installedModMap[dependency.uniqueID]
+        if (entries.size == 0) {
+            for (key in installedModMap.keySet()) {
+                if (StringUtils.equalsIgnoreCase(key, dependency.uniqueID)) {
+                    dependency.uniqueID = key
+                    entries = installedModMap[dependency.uniqueID]
+                    break
                 }
             }
         }
-        if (entries.size() != 1) {
-            return true;
+        if (entries.size != 1) {
+            return true
         }
-        String version = entries.get(0).getVersion();
+        val version = entries[0].version
         if (StringUtils.isBlank(version)) {
-            return true;
+            return true
         }
-        if (StringUtils.isBlank(dependency.getMinimumVersion())) {
-            return false;
-        }
-        return VersionUtil.compareVersion(version, dependency.getMinimumVersion()) < 0;
+        return if (StringUtils.isBlank(dependency.minimumVersion)) {
+            false
+        } else VersionUtil.compareVersion(version, dependency.minimumVersion) < 0
     }
 
-    private Tuple2<String, String> checkContentPackDependencyError(ModManifestEntry mod, ImmutableListMultimap<String, ModManifestEntry> installedModMap) {
-        ModManifestEntry dependency = mod.getContentPackFor();
+    private fun checkContentPackDependencyError(
+        mod: ModManifestEntry, installedModMap: ImmutableListMultimap<String, ModManifestEntry>
+    ): Tuple2<String?, String?>? {
+        val dependency = mod.contentPackFor
         if (dependency != null) {
-            if (dependency.getIsRequired() != null && !dependency.getIsRequired()) {
-                return null;
+            if (dependency.isRequired != null && !dependency.isRequired) {
+                return null
             }
-            ImmutableList<ModManifestEntry> entries = installedModMap.get(dependency.getUniqueID());
-            if (entries.size() == 0) {
-                for (String key : installedModMap.keySet()) {
-                    if (StringUtils.equalsIgnoreCase(key, dependency.getUniqueID())) {
-                        dependency.setUniqueID(key);
-                        entries = installedModMap.get(dependency.getUniqueID());
-                        break;
+            var entries = installedModMap[dependency.uniqueID]
+            if (entries.size == 0) {
+                for (key in installedModMap.keySet()) {
+                    if (StringUtils.equalsIgnoreCase(key, dependency.uniqueID)) {
+                        dependency.uniqueID = key
+                        entries = installedModMap[dependency.uniqueID]
+                        break
                     }
                 }
             }
-            if (entries.size() != 1) {
-                return new Tuple2<>(root.getContext().getString(R.string.error_depends_on_mod, mod.getUniqueID(), dependency.getUniqueID()), dependency.getUniqueID());
+            if (entries.size != 1) {
+                return Tuple2(
+                    root.context.getString(
+                        R.string.error_depends_on_mod, mod.uniqueID, dependency.uniqueID
+                    ), dependency.uniqueID
+                )
             }
-            String version = entries.get(0).getVersion();
+            val version = entries[0].version
             if (!StringUtils.isBlank(version)) {
-                if (StringUtils.isBlank(dependency.getMinimumVersion())) {
-                    return null;
+                if (StringUtils.isBlank(dependency.minimumVersion)) {
+                    return null
                 }
-                if (VersionUtil.compareVersion(version, dependency.getMinimumVersion()) < 0) {
-                    return new Tuple2<>(root.getContext().getString(R.string.error_depends_on_mod_version, mod.getUniqueID(), dependency.getUniqueID(), dependency.getMinimumVersion()), dependency.getUniqueID());
+                if (VersionUtil.compareVersion(version, dependency.minimumVersion) < 0) {
+                    return Tuple2(
+                        root.context.getString(
+                            R.string.error_depends_on_mod_version,
+                            mod.uniqueID,
+                            dependency.uniqueID,
+                            dependency.minimumVersion
+                        ), dependency.uniqueID
+                    )
                 }
             }
-            return null;
+            return null
         }
-        return null;
+        return null
+    }
+
+    companion object {
+        private const val TAG = "MANAGER"
+
+        /**
+         * 查找第一个匹配的Mod
+         *
+         * @param filter 过滤规则
+         * @return Mod信息
+         */
+        @JvmStatic
+        fun findFirstModIf(filter: (ModManifestEntry) -> Boolean): ModManifestEntry? {
+            val files = Queues.newConcurrentLinkedQueue<File>()
+            files.add(File(FileUtils.stadewValleyBasePath, Constants.MOD_PATH))
+            do {
+                val currentFile = files.poll()
+                if (currentFile != null && currentFile.exists()) {
+                    var foundManifest = false
+                    val listFiles = currentFile.listFiles { obj -> obj.isFile } ?: continue
+                    for (file in listFiles) {
+                        if (StringUtils.equalsIgnoreCase(file.name, "manifest.json")) {
+                            val manifest = FileUtils.getFileJson(file, ModManifestEntry::class.java)
+                            foundManifest = true
+                            if (manifest != null) {
+                                manifest.assetPath = file.parentFile?.absolutePath
+                                if (filter.invoke(manifest)) {
+                                    return manifest
+                                }
+                            }
+                            break
+                        }
+                    }
+                    if (!foundManifest) {
+                        val subFolder = currentFile.listFiles { obj -> obj.isDirectory }
+                        subFolder?.let { files.addAll(Lists.newArrayList(*it)) }
+                    }
+                }
+            } while (!files.isEmpty())
+            return null
+        }
+
+        /**
+         * 查找全部已识别Mod
+         *
+         * @param ignoreDisabledMod 是否忽略禁用的mod
+         * @return Mod信息列表
+         */
+        @JvmOverloads
+        fun findAllInstalledMods(ignoreDisabledMod: Boolean = false): MutableList<ModManifestEntry> {
+            val files = Queues.newConcurrentLinkedQueue<File>()
+            files.add(File(FileUtils.stadewValleyBasePath, Constants.MOD_PATH))
+            val mods: MutableList<ModManifestEntry> = ArrayList(30)
+            do {
+                val currentFile = files.poll()
+                if (currentFile != null && currentFile.exists()) {
+                    var foundManifest = false
+                    val listFiles = currentFile.listFiles { obj -> obj.isFile }
+                    if (listFiles != null) {
+                        for (file in listFiles) {
+                            if (StringUtils.equalsIgnoreCase(file.name, "manifest.json")) {
+                                val manifest =
+                                    FileUtils.getFileJson(file, ModManifestEntry::class.java)
+                                foundManifest = true
+                                if (manifest != null && StringUtils.isNoneBlank(manifest.uniqueID)) {
+                                    if (ignoreDisabledMod && StringUtils.startsWith(
+                                            file.parentFile?.name, "."
+                                        )
+                                    ) {
+                                        break
+                                    }
+                                    manifest.assetPath = file.parentFile?.absolutePath
+                                    manifest.lastModified = file.lastModified()
+                                    mods.add(manifest)
+                                }
+                                break
+                            }
+                        }
+                    }
+                    if (!foundManifest) {
+                        val listDirectories = currentFile.listFiles { obj -> obj.isDirectory }
+                        if (listDirectories != null) {
+                            files.addAll(Lists.newArrayList(*listDirectories))
+                        }
+                    }
+                }
+            } while (!files.isEmpty())
+            return mods
+        }
     }
 }
