@@ -1,16 +1,12 @@
-package com.zane.smapiinstaller.constant;
+package com.zane.smapiinstaller.constant
 
 /**
  * @author Zane
  */
-public class AppConfigKeyConstants {
-    public static final String ACTIVE_TRANSLATOR = "ActiveTranslator";
-
-    public static final String MOD_LIST_SORT_BY = "ModListSortBy";
-
-    public static final String IGNORE_UPDATE_VERSION_CODE = "UpdateIgnoreVersionCode";
-
-    public static final String PRIVACY_POLICY_CONFIRM = "PrivacyPolicyConfirm";
-
-    public static final String ADVANCED_MODE = "AdvancedMode";
+object AppConfigKeyConstants {
+    const val ACTIVE_TRANSLATOR = "ActiveTranslator"
+    const val MOD_LIST_SORT_BY = "ModListSortBy"
+    const val IGNORE_UPDATE_VERSION_CODE = "UpdateIgnoreVersionCode"
+    const val PRIVACY_POLICY_CONFIRM = "PrivacyPolicyConfirm"
+    const val ADVANCED_MODE = "AdvancedMode"
 }
