@@ -1,10 +1,9 @@
-package com.zane.smapiinstaller.constant;
+package com.zane.smapiinstaller.constant
 
 /**
  * @author Zane
  */
-public class DownloadableContentTypeConstants {
-    public static final String LOCALE = "LOCALE";
-
-    public static final String COMPAT = "COMPAT";
+object DownloadableContentTypeConstants {
+    const val LOCALE = "LOCALE"
+    const val COMPAT = "COMPAT"
 }
