@@ -1,20 +1,15 @@
-package com.zane.smapiinstaller.constant;
+package com.zane.smapiinstaller.constant
 
 /**
  * @author Zane
  */
-public class ManifestPatchConstants {
-    public static final String APP_NAME = "Stardew Valley";
-
-    public static final String PATTERN_MAIN_ACTIVITY = ".MainActivity";
-
-    public static final String PATTERN_VERSION_CODE = "versionCode";
-
-    public static final String PATTERN_VERSION_NAME = "versionName";
-
-    public static final CharSequence APP_PACKAGE_NAME = "com.chucklefish.stardewvalley";
-
-    public static final String PATTERN_VERSION_AMAZON = "amazon";
-
-    public static final String PATTERN_EXTRACT_NATIVE_LIBS = "extractNativeLibs";
+object ManifestPatchConstants {
+    const val APP_NAME = "Stardew Valley"
+    const val PATTERN_MAIN_ACTIVITY = ".MainActivity"
+    const val PATTERN_VERSION_CODE = "versionCode"
+    const val PATTERN_VERSION_NAME = "versionName"
+    @JvmField
+    val APP_PACKAGE_NAME = "com.chucklefish.stardewvalley"
+    const val PATTERN_VERSION_AMAZON = "amazon"
+    const val PATTERN_EXTRACT_NATIVE_LIBS = "extractNativeLibs"
 }
