@@ -1,97 +1,80 @@
-package com.zane.smapiinstaller.logic;
+package com.zane.smapiinstaller.logic
 
-import android.app.Activity;
-import android.content.ActivityNotFoundException;
-import android.content.ClipData;
-import android.content.ClipboardManager;
-import android.content.ContentResolver;
-import android.content.Context;
-import android.content.ContextWrapper;
-import android.content.Intent;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
-import android.net.Uri;
-import android.os.Build;
-import android.provider.DocumentsContract;
-import android.provider.Settings;
-import android.util.Log;
-import android.view.View;
-import android.view.animation.Animation;
-import android.view.animation.AnimationUtils;
-import android.widget.ImageView;
-
-import androidx.annotation.RequiresApi;
-import androidx.documentfile.provider.DocumentFile;
-import androidx.documentfile.provider.DocumentUtils;
-
-import com.afollestad.materialdialogs.MaterialDialog;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Lists;
-import com.google.common.io.ByteStreams;
-import com.lmntrx.android.library.livin.missme.ProgressDialog;
-import com.microsoft.appcenter.crashes.Crashes;
-import com.zane.smapiinstaller.MainApplication;
-import com.zane.smapiinstaller.R;
-import com.zane.smapiinstaller.constant.Constants;
-import com.zane.smapiinstaller.constant.DialogAction;
-import com.zane.smapiinstaller.constant.ManifestPatchConstants;
-import com.zane.smapiinstaller.entity.ApkFilesManifest;
-import com.zane.smapiinstaller.entity.ManifestEntry;
-import com.zane.smapiinstaller.utils.DialogUtils;
-import com.zane.smapiinstaller.utils.FileUtils;
-import com.zane.smapiinstaller.utils.StringUtils;
-import com.zane.smapiinstaller.utils.ZipUtils;
-
-import org.apache.commons.io.FilenameUtils;
-import org.apache.commons.io.filefilter.WildcardFileFilter;
-import org.zeroturnaround.zip.ZipUtil;
-
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.FileFilter;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.nio.channels.Channels;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.BiConsumer;
-import java.util.function.Consumer;
-import java.util.function.Function;
-import java.util.stream.Stream;
-
-import pxb.android.axml.AxmlReader;
-import pxb.android.axml.AxmlVisitor;
-import pxb.android.axml.AxmlWriter;
-import pxb.android.axml.NodeVisitor;
+import android.app.Activity
+import android.content.ActivityNotFoundException
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.ContentResolver
+import android.content.Context
+import android.content.ContextWrapper
+import android.content.Intent
+import android.content.pm.PackageInfo
+import android.net.Uri
+import android.os.Build
+import android.provider.DocumentsContract
+import android.provider.Settings
+import android.util.Log
+import android.view.View
+import android.view.animation.Animation
+import android.view.animation.AnimationUtils
+import android.widget.ImageView
+import androidx.annotation.RequiresApi
+import androidx.documentfile.provider.DocumentFile
+import androidx.documentfile.provider.DocumentUtils
+import com.afollestad.materialdialogs.MaterialDialog
+import com.fasterxml.jackson.core.type.TypeReference
+import com.google.common.collect.Iterables
+import com.google.common.collect.Lists
+import com.google.common.io.ByteStreams
+import com.lmntrx.android.library.livin.missme.ProgressDialog
+import com.microsoft.appcenter.crashes.Crashes
+import com.zane.smapiinstaller.MainApplication
+import com.zane.smapiinstaller.R
+import com.zane.smapiinstaller.constant.Constants
+import com.zane.smapiinstaller.constant.DialogAction
+import com.zane.smapiinstaller.constant.ManifestPatchConstants
+import com.zane.smapiinstaller.entity.ApkFilesManifest
+import com.zane.smapiinstaller.entity.ManifestEntry
+import com.zane.smapiinstaller.logic.ActivityResultHandler.registerListener
+import com.zane.smapiinstaller.utils.DialogUtils
+import com.zane.smapiinstaller.utils.FileUtils
+import com.zane.smapiinstaller.utils.StringUtils
+import com.zane.smapiinstaller.utils.ZipUtils
+import org.apache.commons.io.FilenameUtils
+import org.apache.commons.io.filefilter.WildcardFileFilter
+import org.zeroturnaround.zip.ZipUtil
+import java.io.ByteArrayInputStream
+import java.io.File
+import java.io.FileFilter
+import java.io.FileOutputStream
+import java.io.IOException
+import java.nio.channels.Channels
+import java.util.Collections
 
 /**
  * 通用逻辑
  *
  * @author Zane
  */
-public class CommonLogic {
+object CommonLogic {
     /**
      * 从View获取所属Activity
      *
      * @param view context容器
      * @return Activity
      */
-    public static Activity getActivityFromView(View view) {
+    @JvmStatic
+    fun getActivityFromView(view: View?): Activity? {
         if (null != view) {
-            Context context = view.getContext();
-            while (context instanceof ContextWrapper) {
-                if (context instanceof Activity) {
-                    return (Activity) context;
+            var context = view.context
+            while (context is ContextWrapper) {
+                if (context is Activity) {
+                    return context
                 }
-                context = ((ContextWrapper) context).getBaseContext();
+                context = context.baseContext
             }
         }
-        return null;
+        return null
     }
 
     /**
@@ -100,25 +83,12 @@ public class CommonLogic {
      * @param view 控件
      * @return Application
      */
-    public static MainApplication getApplicationFromView(View view) {
-        Activity activity = getActivityFromView(view);
-        if (null != activity) {
-            return (MainApplication) activity.getApplication();
-        }
-        return null;
-    }
-
-    /**
-     * 当data非null时执行操作
-     *
-     * @param data   数据
-     * @param action 操作
-     * @param <T>    泛型
-     */
-    public static <T> void doOnNonNull(T data, Consumer<T> action) {
-        if (data != null) {
-            action.accept(data);
-        }
+    @JvmStatic
+    fun getApplicationFromView(view: View?): MainApplication? {
+        val activity = getActivityFromView(view)
+        return if (null != activity) {
+            activity.application as MainApplication
+        } else null
     }
 
     /**
@@ -127,9 +97,12 @@ public class CommonLogic {
      * @param activity activity
      * @param action   操作
      */
-    public static void runOnUiThread(Activity activity, Consumer<Activity> action) {
-        if (activity != null && !activity.isFinishing()) {
-            activity.runOnUiThread(() -> action.accept(activity));
+    @JvmStatic
+    fun runOnUiThread(activity: Activity?, action: (Activity) -> Unit) {
+        activity?.let {
+            if (!it.isFinishing) {
+                it.runOnUiThread { action.invoke(it) }
+            }
         }
     }
 
@@ -139,13 +112,14 @@ public class CommonLogic {
      * @param context context
      * @param url     目标URL
      */
-    public static void openUrl(Context context, String url) {
+    @JvmStatic
+    fun openUrl(context: Context, url: String?) {
         try {
-            Intent intent = new Intent();
-            intent.setData(Uri.parse(url));
-            intent.setAction(Intent.ACTION_VIEW);
-            context.startActivity(intent);
-        } catch (ActivityNotFoundException ignored) {
+            val intent = Intent()
+            intent.data = Uri.parse(url)
+            intent.action = Intent.ACTION_VIEW
+            context.startActivity(intent)
+        } catch (ignored: ActivityNotFoundException) {
         }
     }
 
@@ -156,15 +130,16 @@ public class CommonLogic {
      * @param copyStr 文本
      * @return 是否复制成功
      */
-    public static boolean copyToClipboard(Context context, String copyStr) {
-        try {
-            CommonLogic.doOnNonNull(context.getSystemService(Context.CLIPBOARD_SERVICE), cm -> {
-                ClipData mClipData = ClipData.newPlainText("Label", copyStr);
-                ((ClipboardManager) cm).setPrimaryClip(mClipData);
-            });
-            return true;
-        } catch (Exception e) {
-            return false;
+    @JvmStatic
+    fun copyToClipboard(context: Context, copyStr: String?): Boolean {
+        return try {
+            context.getSystemService(Context.CLIPBOARD_SERVICE)?.let { cm ->
+                val mClipData = ClipData.newPlainText("Label", copyStr)
+                (cm as ClipboardManager).setPrimaryClip(mClipData)
+            }
+            true
+        } catch (e: Exception) {
+            false
         }
     }
 
@@ -174,38 +149,48 @@ public class CommonLogic {
      * @param context context
      * @return 兼容包列表
      */
-    public static List<ApkFilesManifest> findAllApkFileManifest(Context context) {
-        ApkFilesManifest apkFilesManifest = FileUtils.getAssetJson(context, "apk_files_manifest.json", ApkFilesManifest.class);
-        ArrayList<ApkFilesManifest> apkFilesManifests = Lists.newArrayList(apkFilesManifest);
-        File compatFolder = new File(context.getFilesDir(), "compat");
+    fun findAllApkFileManifest(context: Context): MutableList<ApkFilesManifest> {
+        val apkFilesManifest =
+            FileUtils.getAssetJson(context, "apk_files_manifest.json", ApkFilesManifest::class.java)
+        val apkFilesManifests =
+            apkFilesManifest?.let { Lists.newArrayList(apkFilesManifest) } ?: Lists.newArrayList()
+        val compatFolder = File(context.filesDir, "compat")
         if (compatFolder.exists()) {
-            for (File directory : compatFolder.listFiles(File::isDirectory)) {
-                File manifestFile = new File(directory, "apk_files_manifest.json");
-                if (manifestFile.exists()) {
-                    ApkFilesManifest manifest = FileUtils.getFileJson(manifestFile, ApkFilesManifest.class);
-                    if (manifest != null) {
-                        apkFilesManifests.add(manifest);
+            compatFolder.listFiles { obj -> obj.isDirectory }?.let {
+                for (directory in it) {
+                    val manifestFile = File(directory, "apk_files_manifest.json")
+                    if (manifestFile.exists()) {
+                        val manifest =
+                            FileUtils.getFileJson(manifestFile, ApkFilesManifest::class.java)
+                        if (manifest != null) {
+                            apkFilesManifests.add(manifest)
+                        }
                     }
                 }
             }
         }
-        Collections.sort(apkFilesManifests, (a, b) -> {
-            if (a.getTargetPackageName() != null && b.getTargetPackageName() == null) {
-                return -1;
-            } else if (b.getTargetPackageName() != null) {
-                return Long.compare(b.getMinBuildCode(), a.getMinBuildCode());
+        apkFilesManifests.sortWith(Comparator sort@{ a, b ->
+            if (a.targetPackageName != null && b.targetPackageName == null) {
+                return@sort -1
+            } else if (b.targetPackageName != null) {
+                return@sort b.minBuildCode.compareTo(a.minBuildCode)
             }
-            return 1;
-        });
-        return apkFilesManifests;
+            1
+        })
+        return apkFilesManifests
     }
 
-    public static String computePackageName(PackageInfo packageInfo) {
-        String packageName = packageInfo.packageName;
-        if (StringUtils.endsWith(packageInfo.versionName, ManifestPatchConstants.PATTERN_VERSION_AMAZON)) {
-            packageName = ManifestPatchConstants.APP_PACKAGE_NAME + ManifestPatchConstants.PATTERN_VERSION_AMAZON;
+    @JvmStatic
+    fun computePackageName(packageInfo: PackageInfo): String {
+        var packageName = packageInfo.packageName
+        if (StringUtils.endsWith(
+                packageInfo.versionName, ManifestPatchConstants.PATTERN_VERSION_AMAZON
+            )
+        ) {
+            packageName =
+                ManifestPatchConstants.APP_PACKAGE_NAME + ManifestPatchConstants.PATTERN_VERSION_AMAZON
         }
-        return packageName;
+        return packageName
     }
 
     /**
@@ -218,238 +203,295 @@ public class CommonLogic {
      * @param versionCode 版本号
      * @return 操作是否成功
      */
-    public static boolean unpackSmapiFiles(Activity context, String apkPath, boolean checkMode, String packageName, long versionCode) {
-        List<ApkFilesManifest> apkFilesManifests = CommonLogic.findAllApkFileManifest(context);
-        filterManifest(apkFilesManifests, packageName, versionCode);
-        List<ManifestEntry> manifestEntries = null;
-        ApkFilesManifest apkFilesManifest = null;
-        if (apkFilesManifests.size() > 0) {
-            apkFilesManifest = apkFilesManifests.get(0);
-            String basePath = apkFilesManifest.getBasePath();
+    @JvmStatic
+    fun unpackSmapiFiles(
+        context: Activity,
+        apkPath: String,
+        checkMode: Boolean,
+        packageName: String?,
+        versionCode: Long
+    ): Boolean {
+        val apkFilesManifests = findAllApkFileManifest(context)
+        filterManifest(apkFilesManifests, packageName, versionCode)
+        var manifestEntries: List<ManifestEntry>? = null
+        var apkFilesManifest: ApkFilesManifest? = null
+        if (apkFilesManifests.size > 0) {
+            apkFilesManifest = apkFilesManifests[0]
+            val basePath = apkFilesManifests[0].basePath
             if (StringUtils.isNoneBlank(basePath)) {
-                manifestEntries = FileUtils.getAssetJson(context, basePath + "smapi_files_manifest.json", new TypeReference<List<ManifestEntry>>() {
-                });
+                manifestEntries = FileUtils.getAssetJson(context,
+                    basePath + "smapi_files_manifest.json",
+                    object : TypeReference<List<ManifestEntry>>() {})
             }
         }
         if (manifestEntries == null) {
-            manifestEntries = FileUtils.getAssetJson(context, "smapi_files_manifest.json", new TypeReference<List<ManifestEntry>>() {
-            });
+            manifestEntries = FileUtils.getAssetJson(context,
+                "smapi_files_manifest.json",
+                object : TypeReference<List<ManifestEntry>>() {})
         }
         if (manifestEntries == null) {
-            return false;
+            return false
         }
-        File basePath = new File(FileUtils.getStadewValleyBasePath() + "/StardewValley/");
+        val basePath = File(FileUtils.stadewValleyBasePath + "/StardewValley/")
         if (!basePath.exists()) {
             if (!basePath.mkdir()) {
-                return false;
+                return false
             }
         } else {
             if (!checkMode) {
-                File[] oldAssemblies = new File(basePath, "smapi-internal").listFiles((FileFilter) new WildcardFileFilter("*.dll"));
+                val oldAssemblies = File(
+                    basePath, "smapi-internal"
+                ).listFiles(WildcardFileFilter("*.dll") as FileFilter)
                 if (oldAssemblies != null) {
-                    for (File file : oldAssemblies) {
-                        FileUtils.deleteQuietly(file);
+                    for (file in oldAssemblies) {
+                        org.zeroturnaround.zip.commons.FileUtils.deleteQuietly(file)
                     }
                 }
             }
         }
-        File noMedia = new File(basePath, ".nomedia");
+        val noMedia = File(basePath, ".nomedia")
         if (!noMedia.exists()) {
             try {
-                noMedia.createNewFile();
-            } catch (IOException ignored) {
+                noMedia.createNewFile()
+            } catch (ignored: IOException) {
             }
         }
-        for (ManifestEntry entry : manifestEntries) {
-            File targetFile = new File(basePath, entry.getTargetPath());
-            switch (entry.getOrigin()) {
-                case 0:
-                    unpackFromInstaller(context, checkMode, apkFilesManifest, basePath, entry, targetFile);
-                    break;
-                case 1:
-                    unpackFromApk(apkPath, checkMode, entry, targetFile);
-                    break;
-                default:
-                    break;
+        for (entry in manifestEntries) {
+            val targetFile = File(basePath, entry.targetPath)
+            when (entry.origin) {
+                0 -> unpackFromInstaller(
+                    context, checkMode, apkFilesManifest, basePath, entry, targetFile
+                )
+
+                1 -> unpackFromApk(apkPath, checkMode, entry, targetFile)
+                else -> {}
             }
         }
-        if (CommonLogic.checkDataRootPermission(context)) {
-            Uri targetDirUri = pathToTreeUri(Constants.TARGET_DATA_FILE_URI);
-            DocumentFile documentFile = DocumentFile.fromTreeUri(context, targetDirUri);
-            DocumentFile filesDoc = DocumentUtils.findFile(context, documentFile, "files");
-            if(filesDoc != null) {
-                copyDocument(context, new File(basePath, "smapi-internal"), filesDoc);
+        if (checkDataRootPermission(context)) {
+            val targetDirUri = pathToTreeUri(Constants.TARGET_DATA_FILE_URI)
+            val documentFile = DocumentFile.fromTreeUri(context, targetDirUri)
+            val filesDoc = DocumentUtils.findFile(context, documentFile, "files")
+            if (filesDoc != null) {
+                copyDocument(context, File(basePath, "smapi-internal"), filesDoc)
             }
         }
-        return true;
+        return true
     }
 
-    public static void copyDocument(Activity context, File src, DocumentFile dest) {
-        if (src.isDirectory()) {
-            DocumentFile documentFile = DocumentUtils.findFile(context, dest, src.getName());
-            if (documentFile == null) {
-                documentFile = dest.createDirectory(src.getName());
-            }
-            for (File file : src.listFiles()) {
-                copyDocument(context, file, documentFile);
+    @JvmStatic
+    fun copyDocument(context: Activity, src: File, dest: DocumentFile) {
+        if (src.isDirectory) {
+            val documentFile =
+                DocumentUtils.findFile(context, dest, src.name) ?: dest.createDirectory(src.name)
+            documentFile?.let {
+                src.listFiles()?.let {
+                    for (file in it) {
+                        copyDocument(context, file, documentFile)
+                    }
+                }
             }
         } else {
-            DocumentFile documentFile = DocumentUtils.findFile(context, dest, src.getName());
-            if (documentFile == null) {
-                documentFile = dest.createFile("application/x-binary", src.getName());
-            }
-            if(documentFile.length() != src.length()) {
-                try (OutputStream outputStream = context.getContentResolver().openOutputStream(documentFile.getUri())) {
-                    FileUtils.copy(src, outputStream);
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
+            val documentFile = DocumentUtils.findFile(context, dest, src.name) ?: dest.createFile(
+                "application/x-binary", src.name
+            )
+            documentFile?.let {
+                if (documentFile.length() != src.length()) {
+                    try {
+                        context.contentResolver.openOutputStream(documentFile.uri)
+                            .use { outputStream ->
+                                org.zeroturnaround.zip.commons.FileUtils.copy(
+                                    src, outputStream
+                                )
+                            }
+                    } catch (e: IOException) {
+                        throw RuntimeException(e)
+                    }
                 }
             }
         }
     }
 
-    private static boolean checkMusic(Context context) {
+    private fun checkMusic(context: Context): Boolean {
         if (Build.VERSION.SDK_INT > Build.VERSION_CODES.Q) {
-            File pathFrom = new File(FileUtils.getStadewValleyBasePath(), "Android/obb/" + Constants.ORIGIN_PACKAGE_NAME_GOOGLE);
-            File pathTo = new File(FileUtils.getStadewValleyBasePath(), "StardewValley");
-            if (pathFrom.exists() && pathFrom.isDirectory()) {
-                if (!checkObbRootPermission((Activity) context, ActivityResultHandler.REQUEST_CODE_OBB_FILES_ACCESS_PERMISSION, (success) -> checkMusic(context))) {
-                    return false;
+            val pathFrom = File(
+                FileUtils.stadewValleyBasePath,
+                "Android/obb/" + Constants.ORIGIN_PACKAGE_NAME_GOOGLE
+            )
+            val pathTo = File(FileUtils.stadewValleyBasePath, "StardewValley")
+            if (pathFrom.exists() && pathFrom.isDirectory) {
+                if (!checkObbRootPermission(
+                        context as Activity,
+                        ActivityResultHandler.REQUEST_CODE_OBB_FILES_ACCESS_PERMISSION
+                    ) { checkMusic(context) }
+                ) {
+                    return false
                 }
                 if (!pathTo.exists()) {
-                    pathTo.mkdirs();
+                    pathTo.mkdirs()
                 }
-                File[] files = pathFrom.listFiles((dir, name) -> name.contains("com.chucklefish.stardewvalley.obb"));
+                val files =
+                    pathFrom.listFiles { _, name -> name.contains("com.chucklefish.stardewvalley.obb") }
                 if (files != null) {
-                    for (File file : files) {
+                    for (file in files) {
                         try {
-                            File targetFile = new File(pathTo, file.getName());
-                            if (!targetFile.exists() || FileUtils.sizeOf(targetFile) != FileUtils.sizeOf(file)) {
-                                FileUtils.copyFile(file, targetFile);
+                            val targetFile = File(pathTo, file.name)
+                            if (!targetFile.exists() || org.zeroturnaround.zip.commons.FileUtils.sizeOf(
+                                    targetFile
+                                ) != org.zeroturnaround.zip.commons.FileUtils.sizeOf(
+                                    file
+                                )
+                            ) {
+                                org.zeroturnaround.zip.commons.FileUtils.copyFile(file, targetFile)
                             }
-                        } catch (IOException e) {
-                            Crashes.trackError(e);
+                        } catch (e: IOException) {
+                            Crashes.trackError(e)
                         }
                     }
                 }
             }
         }
-        return true;
+        return true
     }
 
-    private static void unpackFromApk(String apkPath, boolean checkMode, ManifestEntry entry, File targetFile) {
+    private fun unpackFromApk(
+        apkPath: String, checkMode: Boolean, entry: ManifestEntry, targetFile: File
+    ) {
         if (!checkMode || !targetFile.exists()) {
-            if (entry.isXALZ()) {
-                byte[] bytes = ZipUtil.unpackEntry(new File(apkPath), entry.getAssetPath());
-                if (entry.isXALZ()) {
-                    bytes = ZipUtils.decompressXALZ(bytes);
+            if (entry.isXALZ) {
+                var bytes = ZipUtil.unpackEntry(File(apkPath), entry.assetPath)
+                if (entry.isXALZ) {
+                    bytes = ZipUtils.decompressXALZ(bytes)
                 }
-                try (FileOutputStream outputStream = FileUtils.openOutputStream(targetFile)) {
-                    ByteStreams.copy(Channels.newChannel(new ByteArrayInputStream(bytes)), outputStream.getChannel());
-                } catch (IOException ignore) {
+                try {
+                    org.zeroturnaround.zip.commons.FileUtils.openOutputStream(targetFile)
+                        .use { outputStream ->
+                            ByteStreams.copy(
+                                Channels.newChannel(ByteArrayInputStream(bytes)),
+                                outputStream.channel
+                            )
+                        }
+                } catch (ignore: IOException) {
                 }
             } else {
-                ZipUtil.unpack(new File(apkPath), targetFile, name -> name.startsWith(entry.getAssetPath()) ? FilenameUtils.getName(name) : null);
+                ZipUtil.unpack(File(apkPath), targetFile) { name ->
+                    if (name.startsWith(
+                            entry.assetPath
+                        )
+                    ) FilenameUtils.getName(name) else null
+                }
             }
         }
     }
 
-    private static void unpackFromInstaller(Context context, boolean checkMode, ApkFilesManifest apkFilesManifest, File basePath, ManifestEntry entry, File targetFile) {
-        if (entry.isExternal() && apkFilesManifest != null) {
-            byte[] bytes = FileUtils.getAssetBytes(context, apkFilesManifest.getBasePath() + entry.getAssetPath());
-            try (FileOutputStream outputStream = new FileOutputStream(targetFile)) {
-                ByteStreams.copy(Channels.newChannel(new ByteArrayInputStream(bytes)), outputStream.getChannel());
-            } catch (IOException ignored) {
+    private fun unpackFromInstaller(
+        context: Context,
+        checkMode: Boolean,
+        apkFilesManifest: ApkFilesManifest?,
+        basePath: File,
+        entry: ManifestEntry,
+        targetFile: File
+    ) {
+        if (entry.isExternal && apkFilesManifest != null) {
+            val bytes =
+                FileUtils.getAssetBytes(context, apkFilesManifest.basePath + entry.assetPath)
+            try {
+                FileOutputStream(targetFile).use { outputStream ->
+                    ByteStreams.copy(
+                        Channels.newChannel(
+                            ByteArrayInputStream(bytes)
+                        ), outputStream.channel
+                    )
+                }
+            } catch (ignored: IOException) {
             }
         } else {
-            if (entry.getTargetPath().endsWith("/") && entry.getAssetPath().contains("*")) {
-                String path = StringUtils.substring(entry.getAssetPath(), 0, StringUtils.lastIndexOf(entry.getAssetPath(), "/"));
-                String pattern = StringUtils.substringAfterLast(entry.getAssetPath(), "/");
+            if (entry.targetPath.endsWith("/") && entry.assetPath.contains("*")) {
+                val path = StringUtils.substring(
+                    entry.assetPath, 0, StringUtils.lastIndexOf(entry.assetPath, "/")
+                )
+                val pattern = StringUtils.substringAfterLast(entry.assetPath, "/")
                 try {
-                    Stream.of(context.getAssets().list(path))
-                            .filter(filename -> StringUtils.wildCardMatch(filename, pattern))
-                            .forEach(filename -> unpackFile(context, checkMode, path + "/" + filename, new File(basePath, entry.getTargetPath() + filename)));
-                } catch (IOException ignored) {
+                    context.assets.list(path)?.filter { filename ->
+                        StringUtils.wildCardMatch(
+                            filename, pattern
+                        )
+                    }?.forEach { filename ->
+                        unpackFile(
+                            context,
+                            checkMode,
+                            "$path/$filename",
+                            File(basePath, entry.targetPath + filename)
+                        )
+                    }
+                } catch (ignored: IOException) {
                 }
             } else {
-                unpackFile(context, checkMode, entry.getAssetPath(), targetFile);
+                unpackFile(context, checkMode, entry.assetPath, targetFile)
             }
         }
     }
 
-    public static void filterManifest(List<ApkFilesManifest> manifests, String packageName, long versionCode) {
-        Iterables.removeIf(manifests, manifest -> {
+    fun filterManifest(
+        manifests: MutableList<ApkFilesManifest>, packageName: String?, versionCode: Long
+    ) {
+        Iterables.removeIf(manifests) { manifest ->
             if (manifest == null) {
-                return true;
+                return@removeIf true
             }
-            if (versionCode < manifest.getMinBuildCode()) {
-                return true;
+            if (versionCode < manifest.minBuildCode) {
+                return@removeIf true
             }
-            if (manifest.getMaxBuildCode() != null) {
-                if (versionCode > manifest.getMaxBuildCode()) {
-                    return true;
+            if (manifest.maxBuildCode != null) {
+                if (versionCode > manifest.maxBuildCode) {
+                    return@removeIf true
                 }
             }
-            return manifest.getTargetPackageName() != null && packageName != null && !manifest.getTargetPackageName().contains(packageName);
-        });
+            manifest.targetPackageName != null && packageName != null && !manifest.targetPackageName.contains(
+                packageName
+            )
+        }
     }
 
-    private static void unpackFile(Context context, boolean checkMode, String assertPath, File targetFile) {
+    private fun unpackFile(
+        context: Context, checkMode: Boolean, assertPath: String, targetFile: File
+    ) {
         if (!checkMode || !targetFile.exists()) {
-            try (InputStream inputStream = context.getAssets().open(assertPath)) {
-                if (!targetFile.getParentFile().exists()) {
-                    if (!targetFile.getParentFile().mkdirs()) {
-                        Log.e("COMMON", "Make dirs error");
-                        return;
+            try {
+                context.assets.open(assertPath).use { inputStream ->
+                    targetFile.parentFile?.let {
+                        if (!it.exists()) {
+                            if (!it.mkdirs()) {
+                                Log.e("COMMON", "Make dirs error")
+                                return
+                            }
+                        }
+                    }
+                    FileOutputStream(targetFile).use { outputStream ->
+                        ByteStreams.copy(
+                            Channels.newChannel(
+                                inputStream
+                            ), outputStream.channel
+                        )
                     }
                 }
-                try (FileOutputStream outputStream = new FileOutputStream(targetFile)) {
-                    ByteStreams.copy(Channels.newChannel(inputStream), outputStream.getChannel());
-                }
-            } catch (IOException e) {
-                Log.e("COMMON", "Copy Error", e);
+            } catch (e: IOException) {
+                Log.e("COMMON", "Copy Error", e)
             }
         }
     }
 
-    /**
-     * 修改AndroidManifest.xml文件
-     *
-     * @param bytes            AndroidManifest.xml文件字符数组
-     * @param attrProcessLogic 处理逻辑
-     * @return 修改后的AndroidManifest.xml文件字符数组
-     * @throws IOException 异常
-     */
-    public static byte[] modifyManifest(byte[] bytes, Function<ManifestTagVisitor.AttrArgs, List<ManifestTagVisitor.AttrArgs>> attrProcessLogic, Function<ManifestTagVisitor.ChildArgs, List<ManifestTagVisitor.ChildArgs>> childProcessLogic) throws IOException {
-        AxmlReader reader = new AxmlReader(bytes);
-        AxmlWriter writer = new AxmlWriter();
-        reader.accept(new AxmlVisitor(writer) {
-            @Override
-            public NodeVisitor child(String ns, String name) {
-                NodeVisitor child = super.child(ns, name);
-                return new ManifestTagVisitor(child, attrProcessLogic, childProcessLogic);
-            }
-        });
-        return writer.toByteArray();
-    }
-
-    public static void showAnimation(ImageView view, int anim, Consumer<Animation> action) {
-        Animation animation = AnimationUtils.loadAnimation(getActivityFromView(view), anim);
-        animation.setAnimationListener(new Animation.AnimationListener() {
-            @Override
-            public void onAnimationStart(Animation animation) {
+    @JvmStatic
+    fun showAnimation(view: ImageView, anim: Int, action: (Animation?) -> Unit) {
+        val animation = AnimationUtils.loadAnimation(getActivityFromView(view), anim)
+        animation.setAnimationListener(object : Animation.AnimationListener {
+            override fun onAnimationStart(animation: Animation) {}
+            override fun onAnimationEnd(animation: Animation) {
+                action.invoke(animation)
             }
 
-            @Override
-            public void onAnimationEnd(Animation animation) {
-                action.accept(animation);
-            }
-
-            @Override
-            public void onAnimationRepeat(Animation animation) {
-            }
-        });
-        view.startAnimation(animation);
+            override fun onAnimationRepeat(animation: Animation) {}
+        })
+        view.startAnimation(animation)
     }
 
     /**
@@ -457,17 +499,16 @@ public class CommonLogic {
      *
      * @return 当前应用的版本号
      */
-    public static long getVersionCode(Activity activity) {
+    fun getVersionCode(activity: Activity): Long {
         try {
-            PackageManager manager = activity.getPackageManager();
-            PackageInfo info = manager.getPackageInfo(activity.getPackageName(), 0);
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-                return info.getLongVersionCode();
-            }
-            return info.versionCode;
-        } catch (Exception ignored) {
+            val manager = activity.packageManager
+            val info = manager.getPackageInfo(activity.packageName, 0)
+            return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                info.longVersionCode
+            } else info.versionCode.toLong()
+        } catch (ignored: Exception) {
         }
-        return Integer.MAX_VALUE;
+        return Int.MAX_VALUE.toLong()
     }
 
     /**
@@ -475,156 +516,195 @@ public class CommonLogic {
      *
      * @param activity activity
      */
-    public static void openInPlayStore(Activity activity) {
-        CommonLogic.doOnNonNull(activity, (context) -> {
+    fun openInPlayStore(activity: Activity) {
+        activity.let { context ->
             try {
-                Intent intent = new Intent("android.intent.action.VIEW");
-                intent.setData(Uri.parse("market://details?id=com.zane.smapiinstaller"));
-                intent.setPackage("com.android.vending");
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                context.startActivity(intent);
-            } catch (Exception ex) {
-                CommonLogic.openUrl(activity, "https://play.google.com/store/apps/details?id=com.zane.smapiinstaller");
+                val intent = Intent("android.intent.action.VIEW")
+                intent.data = Uri.parse("market://details?id=com.zane.smapiinstaller")
+                intent.setPackage("com.android.vending")
+                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+                context.startActivity(intent)
+            } catch (ex: Exception) {
+                openUrl(
+                    activity,
+                    "https://play.google.com/store/apps/details?id=com.zane.smapiinstaller"
+                )
             }
-        });
+        }
     }
 
     @RequiresApi(api = Build.VERSION_CODES.R)
-    public static void openPermissionSetting(Activity activity) {
+    fun openPermissionSetting(activity: Activity) {
         try {
-            Intent intent = new Intent(Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION);
-            intent.addCategory("android.intent.category.DEFAULT");
-            intent.setData(Uri.parse("package:" + activity.getPackageName()));
-            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            activity.startActivityForResult(intent, ActivityResultHandler.REQUEST_CODE_ALL_FILES_ACCESS_PERMISSION);
-        } catch (ActivityNotFoundException ignored) {
-            Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-            intent.setData(Uri.parse("package:" + activity.getPackageName()));
-            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            val intent = Intent(Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION)
+            intent.addCategory("android.intent.category.DEFAULT")
+            intent.data = Uri.parse("package:" + activity.packageName)
+            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+            activity.startActivityForResult(
+                intent, ActivityResultHandler.REQUEST_CODE_ALL_FILES_ACCESS_PERMISSION
+            )
+        } catch (ignored: ActivityNotFoundException) {
+            var intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
+            intent.data = Uri.parse("package:" + activity.packageName)
+            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
             try {
-                activity.startActivityForResult(intent, ActivityResultHandler.REQUEST_CODE_ALL_FILES_ACCESS_PERMISSION);
-            } catch (ActivityNotFoundException ignored2) {
-                intent = new Intent(Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION);
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                activity.startActivityForResult(intent, ActivityResultHandler.REQUEST_CODE_ALL_FILES_ACCESS_PERMISSION);
+                activity.startActivityForResult(
+                    intent, ActivityResultHandler.REQUEST_CODE_ALL_FILES_ACCESS_PERMISSION
+                )
+            } catch (ignored2: ActivityNotFoundException) {
+                intent = Intent(Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION)
+                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+                activity.startActivityForResult(
+                    intent, ActivityResultHandler.REQUEST_CODE_ALL_FILES_ACCESS_PERMISSION
+                )
             }
         }
     }
 
-    public static boolean checkDataRootPermission(Context context) {
-        File pathFrom = new File(FileUtils.getStadewValleyBasePath(), "Android/data/" + Constants.TARGET_PACKAGE_NAME + "/files/");
+    @JvmStatic
+    fun checkDataRootPermission(context: Context): Boolean {
+        val pathFrom = File(
+            FileUtils.stadewValleyBasePath,
+            "Android/data/" + Constants.TARGET_PACKAGE_NAME + "/files/"
+        )
         if (!pathFrom.exists()) {
-            return false;
+            return false
         }
-        Uri targetDirUri = pathToTreeUri(Constants.TARGET_DATA_FILE_URI);
-        if (checkPathPermission(context, targetDirUri)) {
-            return true;
-        }
-        return false;
+        val targetDirUri = pathToTreeUri(Constants.TARGET_DATA_FILE_URI)
+        return checkPathPermission(context, targetDirUri)
     }
 
-    public static boolean requestDataRootPermission(Activity context, int REQUEST_CODE_FOR_DIR, Consumer<Boolean> callback) {
-        File pathFrom = new File(FileUtils.getStadewValleyBasePath(), "Android/data/" + Constants.TARGET_PACKAGE_NAME + "/files");
+    @JvmStatic
+    fun requestDataRootPermission(
+        context: Activity, REQUEST_CODE_FOR_DIR: Int, callback: (Boolean) -> Unit
+    ): Boolean {
+        val pathFrom = File(
+            FileUtils.stadewValleyBasePath,
+            "Android/data/" + Constants.TARGET_PACKAGE_NAME + "/files"
+        )
         if (!pathFrom.exists()) {
-            return true;
+            return true
         }
-        Uri targetDirUri = pathToTreeUri(Constants.TARGET_DATA_FILE_URI);
+        val targetDirUri = pathToTreeUri(Constants.TARGET_DATA_FILE_URI)
         if (checkPathPermission(context, targetDirUri)) {
-            return true;
+            return true
         }
-        ActivityResultHandler.registerListener(ActivityResultHandler.REQUEST_CODE_DATA_FILES_ACCESS_PERMISSION, (resultCode, data) -> {
-            takePermission(resultCode, data, context.getContentResolver(), callback);
-        });
-        DocumentFile documentFile = DocumentFile.fromTreeUri(context, targetDirUri);
-        Intent intent1 = new Intent(Intent.ACTION_OPEN_DOCUMENT_TREE);
-        intent1.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION
-                | Intent.FLAG_GRANT_WRITE_URI_PERMISSION
-                | Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION
-                | Intent.FLAG_GRANT_PREFIX_URI_PERMISSION);
-        intent1.putExtra(DocumentsContract.EXTRA_INITIAL_URI, documentFile.getUri());
-        context.startActivityForResult(intent1, REQUEST_CODE_FOR_DIR);
-        return false;
+        registerListener(ActivityResultHandler.REQUEST_CODE_DATA_FILES_ACCESS_PERMISSION) { resultCode, data ->
+            takePermission(
+                resultCode, data, context.contentResolver, callback
+            )
+        }
+        val documentFile = DocumentFile.fromTreeUri(context, targetDirUri)
+        val intent1 = Intent(Intent.ACTION_OPEN_DOCUMENT_TREE)
+        intent1.flags =
+            (Intent.FLAG_GRANT_READ_URI_PERMISSION or Intent.FLAG_GRANT_WRITE_URI_PERMISSION or Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION or Intent.FLAG_GRANT_PREFIX_URI_PERMISSION)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            intent1.putExtra(DocumentsContract.EXTRA_INITIAL_URI, documentFile?.uri)
+        }
+        context.startActivityForResult(intent1, REQUEST_CODE_FOR_DIR)
+        return false
     }
 
-    private static void takePermission(Integer resultCode, Intent data, ContentResolver context, Consumer<Boolean> callback) {
+    private fun takePermission(
+        resultCode: Int, data: Intent?, context: ContentResolver, callback: (Boolean) -> Unit
+    ) {
         if (resultCode == Activity.RESULT_OK) {
             if (data == null) {
-                return;
+                return
             }
-            Uri uri = data.getData();
-            if (uri == null) {
-                return;
-            }
-            context.takePersistableUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION | Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
-            callback.accept(true);
+            val uri = data.data ?: return
+            context.takePersistableUriPermission(
+                uri, Intent.FLAG_GRANT_READ_URI_PERMISSION or Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+            )
+            callback.invoke(true)
         } else {
-            callback.accept(false);
+            callback.invoke(false)
         }
     }
 
-    public static Uri pathToTreeUri(String path) {
-        return Uri.parse("content://com.android.externalstorage.documents/tree/primary%3A" + path.replace("/", "%2F"));
+    @JvmStatic
+    fun pathToTreeUri(path: String): Uri {
+        return Uri.parse(
+            "content://com.android.externalstorage.documents/tree/primary%3A" + path.replace(
+                "/", "%2F"
+            )
+        )
     }
 
-    public static Uri pathToSingleUri(String path) {
-        return Uri.parse("content://com.android.externalstorage.documents/document/primary%3A" + path.replace("/", "%2F"));
+    fun pathToSingleUri(path: String): Uri {
+        return Uri.parse(
+            "content://com.android.externalstorage.documents/document/primary%3A" + path.replace(
+                "/", "%2F"
+            )
+        )
     }
 
-    public static boolean checkPathPermission(Context context, Uri targetDirUri) {
-        if (DocumentFile.fromTreeUri(context, targetDirUri).canWrite()) {
-            return true;
-        }
-        return false;
+    fun checkPathPermission(context: Context, targetDirUri: Uri): Boolean {
+        return DocumentFile.fromTreeUri(context, targetDirUri)?.canWrite() ?: false
     }
 
-    public static boolean checkObbRootPermission(Activity context, int REQUEST_CODE_FOR_DIR, Consumer<Boolean> callback) {
-        Uri targetDirUri = pathToTreeUri("Android/obb");
+    fun checkObbRootPermission(
+        context: Activity, REQUEST_CODE_FOR_DIR: Int, callback: (Boolean) -> Unit
+    ): Boolean {
+        val targetDirUri = pathToTreeUri("Android/obb")
         if (checkPathPermission(context, targetDirUri)) {
-            return true;
+            return true
         }
-        ActivityResultHandler.registerListener(ActivityResultHandler.REQUEST_CODE_OBB_FILES_ACCESS_PERMISSION, (resultCode, data) -> {
-            takePermission(resultCode, data, context.getContentResolver(), callback);
-        });
-        DocumentFile documentFile = DocumentFile.fromTreeUri(context, targetDirUri);
-        Intent intent1 = new Intent(Intent.ACTION_OPEN_DOCUMENT_TREE);
-        intent1.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION
-                | Intent.FLAG_GRANT_WRITE_URI_PERMISSION
-                | Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION
-                | Intent.FLAG_GRANT_PREFIX_URI_PERMISSION);
-        intent1.putExtra(DocumentsContract.EXTRA_INITIAL_URI, documentFile.getUri());
-        context.startActivityForResult(intent1, REQUEST_CODE_FOR_DIR);
-        return false;
+        registerListener(ActivityResultHandler.REQUEST_CODE_OBB_FILES_ACCESS_PERMISSION) { resultCode, data ->
+            takePermission(
+                resultCode, data, context.contentResolver, callback
+            )
+        }
+        val documentFile = DocumentFile.fromTreeUri(context, targetDirUri)
+        val intent1 = Intent(Intent.ACTION_OPEN_DOCUMENT_TREE)
+        intent1.flags =
+            (Intent.FLAG_GRANT_READ_URI_PERMISSION or Intent.FLAG_GRANT_WRITE_URI_PERMISSION or Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION or Intent.FLAG_GRANT_PREFIX_URI_PERMISSION)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            intent1.putExtra(DocumentsContract.EXTRA_INITIAL_URI, documentFile?.uri)
+        }
+        context.startActivityForResult(intent1, REQUEST_CODE_FOR_DIR)
+        return false
     }
 
-    public static void showPrivacyPolicy(View view, BiConsumer<MaterialDialog, DialogAction> callback) {
-        Context context = view.getContext();
-        String policy = FileUtils.getLocaledAssetText(context, "privacy_policy.txt");
-        DialogUtils.showConfirmDialog(view, R.string.privacy_policy, policy, R.string.confirm, R.string.cancel, true, callback);
+    @JvmStatic
+    fun showPrivacyPolicy(
+        view: View, callback: (MaterialDialog?, DialogAction) -> Unit = { _, _ -> }
+    ) {
+        val context = view.context
+        val policy = FileUtils.getLocaledAssetText(context, "privacy_policy.txt")
+        DialogUtils.showConfirmDialog(
+            view, R.string.privacy_policy, policy, R.string.confirm, R.string.cancel, true, callback
+        )
     }
 
-    public static void showProgressDialog(View root, Context context, Consumer<ProgressDialog> dialogConsumer) {
-        AtomicReference<ProgressDialog> dialogHolder = DialogUtils.showProgressDialog(root, R.string.install_progress_title, context.getString(R.string.extracting_package));
-        ProgressDialog dialog = null;
+    @JvmStatic
+    fun showProgressDialog(
+        root: View?, context: Context, dialogConsumer: (ProgressDialog) -> Unit
+    ) {
+        val dialogHolder = DialogUtils.showProgressDialog(
+            root, R.string.install_progress_title, context.getString(R.string.extracting_package)
+        )
+        var dialog: ProgressDialog? = null
         try {
             do {
-                Thread.sleep(10);
-                dialog = dialogHolder.get();
-            } while (dialog == null);
+                Thread.sleep(10)
+                dialog = dialogHolder.get()
+            } while (dialog == null)
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
-                ConfigManager configManager = new ConfigManager();
-                if (configManager.getConfig().isInitial()) {
-                    configManager.getConfig().setInitial(false);
-                    configManager.getConfig().setDisableMonoMod(true);
-                    configManager.flushConfig();
+                val configManager = ConfigManager()
+                if (configManager.config.isInitial) {
+                    configManager.config.isInitial = false
+                    configManager.config.isDisableMonoMod = true
+                    configManager.flushConfig()
                 }
             }
-            dialogConsumer.accept(dialog);
-        } catch (InterruptedException ignored) {
-        } catch (Exception e) {
-            Crashes.trackError(e);
-            DialogUtils.showAlertDialog(root, R.string.error, e.getLocalizedMessage());
+            dialogConsumer.invoke(dialog)
+        } catch (ignored: InterruptedException) {
+        } catch (e: Exception) {
+            Crashes.trackError(e)
+            DialogUtils.showAlertDialog(root, R.string.error, e.localizedMessage)
         } finally {
-            DialogUtils.dismissDialog(root, dialog);
+            DialogUtils.dismissDialog(root, dialog)
         }
     }
 }
