@@ -1,75 +1,72 @@
-package com.zane.smapiinstaller.ui.main;
+package com.zane.smapiinstaller.ui.main
 
-import android.os.Bundle;
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.zane.smapiinstaller.R
+import com.zane.smapiinstaller.ui.about.AboutFragment
+import com.zane.smapiinstaller.ui.config.ConfigFragment
+import com.zane.smapiinstaller.ui.download.DownloadableContentFragment
+import com.zane.smapiinstaller.ui.help.HelpFragment
+import com.zane.smapiinstaller.ui.install.InstallFragment
 
-import com.zane.smapiinstaller.R;
-import com.zane.smapiinstaller.ui.about.AboutFragment;
-import com.zane.smapiinstaller.ui.config.ConfigFragment;
-import com.zane.smapiinstaller.ui.download.DownloadableContentFragment;
-import com.zane.smapiinstaller.ui.help.HelpFragment;
-import com.zane.smapiinstaller.ui.install.InstallFragment;
+internal class MainTabPagerAdapter(private val fragment: Fragment) : FragmentStateAdapter(
+    fragment
+) {
+    override fun createFragment(position: Int): Fragment {
+        val args = Bundle()
+        val fragment: Fragment
+        return when (position) {
+            1 -> {
+                fragment = ConfigFragment()
+                fragment.setArguments(args)
+                fragment
+            }
 
-import org.jetbrains.annotations.NotNull;
+            2 -> {
+                fragment = DownloadableContentFragment()
+                fragment.setArguments(args)
+                fragment
+            }
 
-import androidx.fragment.app.Fragment;
-import androidx.viewpager2.adapter.FragmentStateAdapter;
+            3 -> {
+                fragment = HelpFragment()
+                fragment.setArguments(args)
+                fragment
+            }
 
-class MainTabPagerAdapter extends FragmentStateAdapter {
-    private final Fragment fragment;
+            4 -> {
+                fragment = AboutFragment()
+                fragment.setArguments(args)
+                fragment
+            }
 
-    public MainTabPagerAdapter(Fragment fragment) {
-        super(fragment);
-        this.fragment = fragment;
-    }
+            0 -> {
+                fragment = InstallFragment()
+                fragment.setArguments(args)
+                fragment
+            }
 
-    @NotNull
-    @Override
-    public Fragment createFragment(int position) {
-        Bundle args = new Bundle();
-        Fragment fragment;
-        switch (position){
-            case 1:
-                fragment = new ConfigFragment();
-                fragment.setArguments(args);
-                return fragment;
-            case 2:
-                fragment = new DownloadableContentFragment();
-                fragment.setArguments(args);
-                return fragment;
-            case 3:
-                fragment = new HelpFragment();
-                fragment.setArguments(args);
-                return fragment;
-            case 4:
-                fragment = new AboutFragment();
-                fragment.setArguments(args);
-                return fragment;
-            case 0:
-            default:
-                fragment = new InstallFragment();
-                fragment.setArguments(args);
-                return fragment;
+            else -> {
+                fragment = InstallFragment()
+                fragment.setArguments(args)
+                fragment
+            }
         }
     }
 
-    public String getTitle(int position) {
-        switch (position){
-            case 1:
-                return fragment.getString(R.string.menu_config);
-            case 2:
-                return fragment.getString(R.string.menu_download);
-            case 3:
-                return fragment.getString(R.string.menu_help);
-            case 4:
-                return fragment.getString(R.string.menu_about);
-            case 0:
-            default:
-                return fragment.getString(R.string.menu_install);
+    fun getTitle(position: Int): String {
+        return when (position) {
+            1 -> fragment.getString(R.string.menu_config)
+            2 -> fragment.getString(R.string.menu_download)
+            3 -> fragment.getString(R.string.menu_help)
+            4 -> fragment.getString(R.string.menu_about)
+            0 -> fragment.getString(R.string.menu_install)
+            else -> fragment.getString(R.string.menu_install)
         }
     }
 
-    @Override
-    public int getItemCount() {
-        return 5;
+    override fun getItemCount(): Int {
+        return 5
     }
 }
